// frontend/src/types/generated/models/TimelineEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../serde_json/JsonValue";
import type { TimelineEntryType } from "./TimelineEntryType";

/**
 * 时间线条目
 */
export type TimelineEntry = { id: string, entryType: TimelineEntryType, timestamp: string, iteration: number | null, title: string, description: string | null, actor: string | null, details: JsonValue | null, };